pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "locllm"
version = "0.1.0"
description = "Local LLM toolkit — chat, image-to-markdown, image generation, and more"
requires-python = ">=3.9"
license = { text = "MIT" }
dependencies = [
    "Pillow>=10.0",
    "pytesseract>=0.3.10",
    "numpy>=1.24",
    "click>=8.1",
    "rich>=13.0",
]

[project.optional-dependencies]
ollama  = ["httpx>=0.25"]
imagine = ["diffusers>=0.27", "transformers>=4.38", "accelerate>=0.28", "torch>=2.2"]
all     = ["httpx>=0.25", "diffusers>=0.27", "transformers>=4.38", "accelerate>=0.28", "torch>=2.2"]

[project.scripts]
locllm = "locllm.cli:cli"

[tool.setuptools.packages.find]
where   = ["."]
include = ["locllm*"]
